Revert "Create schema file to contain the schemas and modules for utilize in any file in views"
This reverts commit 0e44156169323eac6e9d39831cdf1cbd00dfd09c.
